refactor(admin): remove mdx editor, change naming convention of admin components,fix constants usage
