--- apps/admin.enes.sh/app/components/InitializedMDXEditor.tsx
@@ -1,56 +1,0 @@
-'use client'
-
-import {
-  AdmonitionDirectiveDescriptor,
-  KitchenSinkToolbar,
-  MDXEditor,
-  MDXEditorMethods,
-  MDXEditorProps,
-  codeBlockPlugin,
-  codeMirrorPlugin,
-  diffSourcePlugin,
-  directivesPlugin,
-  headingsPlugin,
-  linkDialogPlugin,
-  linkPlugin,
-  listsPlugin,
-  markdownShortcutPlugin,
-  quotePlugin,
-  tablePlugin,
-  thematicBreakPlugin,
-  toolbarPlugin
-} from '@mdxeditor/editor'
-import { ForwardedRef } from 'react'
-
-const InitializedMDXEditor = ({
-  editorRef,
-  ...props
-}: { editorRef: ForwardedRef<MDXEditorMethods> | null } & MDXEditorProps) => {
-  return (
-    <MDXEditor
-      ref={editorRef}
-      {...props}
-      plugins={[
-        toolbarPlugin({
-          toolbarContents: () => <KitchenSinkToolbar />
-        }),
-        listsPlugin(),
-        quotePlugin(),
-        headingsPlugin(),
-        linkPlugin(),
-        linkDialogPlugin(),
-        tablePlugin(),
-        thematicBreakPlugin(),
-        codeBlockPlugin({ defaultCodeBlockLanguage: 'txt' }),
-        codeMirrorPlugin({
-          codeBlockLanguages: { js: 'JavaScript', css: 'CSS', txt: 'text', tsx: 'TypeScript' }
-        }),
-        directivesPlugin({ directiveDescriptors: [AdmonitionDirectiveDescriptor] }),
-        diffSourcePlugin({ viewMode: 'rich-text', diffMarkdown: props.markdown }),
-        markdownShortcutPlugin()
-      ]}
-    />
-  )
-}
-
-export default InitializedMDXEditor
