+++ apps/admin.enes.sh/app/components/index.ts
@@ -1,7 +1,5 @@
-export * from './CollapseMenuButton'
-export * from './DataTableColumnHeader'
-export * from './DataTablePagination'
-export * from './ForwardRefEditor'
-export * from './Menu'
-export * from './RedirectButton'
-export * from './SidebarToggle'
+export * from './collapse-menu-button'
+export * from './datatable-column-header'
+export * from './datatable-pagination'
+export * from './menu'
+export * from './sidebar-toggle'
